requirements.txt
asgiref==3.3.0
Django==3.1.3
django-filter==2.4.0
python-decouple==3.3
pytz==2020.4
sqlparse==0.4.1
